Added piezo.c example. pwmWrite() now treats every PWM value >= 4096 as full-on
Added piezo.c example which plays a tune on a piezo ceramic disk if
connected to a PWM pin.
Moved max PWM value of pwmWrite() from 4095 to 4096 since I realized
this is kind of a tri-state. Everything equal or below 0 is full-off,
everything equal or higher than 4096 is full-on and everything in
between (which are values representable by 12 bits [0xFFF]) triggers PWM
signals.